Move fuzzers back to pkg/datapath/linux/route. Isolate fuzzer in new ns
